♻️ clean up, unify and move features to recordService.ts
