+++ src/services/recordService.ts
@@ -12,8 +12,9 @@ import fhirValidator from '../lib/fhirValidator';
 import taggingUtils from '../lib/taggingUtils';
 import documentRoutes from '../routes/documentRoutes';
 import createCryptoService from './createCryptoService';
-import { DecryptedAppData, DecryptedFhirRecord, Key, QueryParams } from './types';
+import { DecryptedAppData, DecryptedFhirRecord, Key, Params, QueryParams } from './types';
 import userService from './userService';
+import { prepareSearchParameters } from './fhirService';
 
 const recordService = {
   updateRecord(
@@ -159,6 +160,46 @@ const recordService = {
       );
   },
 
+  searchWithFallbackIfNeeded(
+    ownerId: string,
+    countOnly: boolean,
+    params: Params = {}
+  ): Promise<any[]> {
+    const searchParams = { ...params };
+
+    const appDataFlag = taggingUtils.generateAppDataFlagTag();
+    if (!params?.tags?.includes(appDataFlag)) {
+      searchParams.exclude_flags = [appDataFlag];
+    }
+
+    const basicParameters = prepareSearchParameters({
+      params: searchParams,
+      fallbackMode: null,
+    });
+
+    if (params?.fhirVersion || params?.annotations || params?.exclude_tags) {
+      /* earlier versions of the Android SDK did not escape the dots in fhir versions,
+      so we query for both of this encoding and the current standard one */
+      const versionFallbackParameters = prepareSearchParameters({
+        params: searchParams,
+        fallbackMode: 'fhirversion',
+      });
+
+      /* original JS SDK implementation was inconsistent in lowercasing/uppercasing
+      some escaped characters, so we query for both versions */
+      const tagFallbackParameters = prepareSearchParameters({
+        params: searchParams,
+        fallbackMode: 'annotation',
+      });
+      return Promise.all([
+        this.searchRecords(ownerId, basicParameters, countOnly),
+        this.searchRecords(ownerId, versionFallbackParameters, countOnly),
+        this.searchRecords(ownerId, tagFallbackParameters, countOnly),
+      ]);
+    }
+    return Promise.all([this.searchRecords(ownerId, basicParameters, countOnly)]);
+  },
+
   searchRecords(ownerId: string, params: QueryParams, countOnly = false): Promise<any> {
     let user;
     let totalCount;
@@ -208,6 +249,36 @@ const recordService = {
           results ? { totalCount, records: reject(results, isError) } : { totalCount }
         )
     );
+  },
+
+  normalizeFallbackSearchResults({responseArray, conversionFunction}) {
+    return responseArray.reduce(
+      (combinedRecords, currentResponse) => {
+        const nonDuplicateRecords = reject(currentResponse.records, newRecord =>
+          combinedRecords.records.some(existingRecord => existingRecord.id === newRecord.id)
+        );
+        const numberOfDuplicates = currentResponse.records.length - nonDuplicateRecords.length;
+        return nonDuplicateRecords?.length
+          ? {
+            records: [
+              ...combinedRecords.records,
+              ...nonDuplicateRecords.map(conversionFunction),
+            ],
+            totalCount:
+              combinedRecords.totalCount +
+              parseInt(currentResponse.totalCount, 10) -
+              numberOfDuplicates,
+          }
+          : {
+            records: [...combinedRecords.records],
+            totalCount: combinedRecords.totalCount,
+          };
+      },
+      {
+        records: [],
+        totalCount: 0,
+      }
+    )
   },
 
   deleteRecord(ownerId: string, recordId: string) {
@@ -263,4 +334,5 @@ const recordService = {
   },
 };
 
+// @ts-ignore
 export default recordService;
